requirements.txt
python-dotenv==1.0.1
slack_bolt==1.22.0
slack_sdk==3.34.0
